second project v1: complete favour part

+++ src/main/java/com/chajiu/service/Impl/FavourServiceImpl.java
@@ -22,6 +22,7 @@ public class FavourServiceImpl implements FavourService {
     public List<Favour> findAllByUserId(Integer uid, Page page) {
         List<Favour> favourList = mapper.findAllByUserId(uid,page);
         for(Favour favour:favourList){
+            favour.getPoetry().setPoetryAuthor(null);   //不返回作者的详细信息
             favour.getPoetry().toSimpleInstance();
         }
         return favourList;
@@ -32,12 +33,15 @@ public class FavourServiceImpl implements FavourService {
         if(mapper.exist(favour)==0)
             mapper.save(favour);
         else
-            throw new Exception("已经收藏！");
+            throw new Exception("诗词编号："+favour.getPoetryId()+" 已经收藏！");
     }
 
     @Override
-    public void delete(Favour favour) {
-        mapper.delete(favour);
+    public void delete(Favour favour) throws Exception {
+        if(mapper.exist(favour)!=0)
+            mapper.delete(favour);
+        else
+            throw new Exception("诗词编号："+favour.getPoetryId()+" 已经取消收藏！");
     }
 
     @Override
